requirements.txt
paho-mqtt>=2.0.0
requests>=2.31.0
pyyaml>=6.0
Pillow>=9.0.0
